Moved dataset to separate file and added default columns
