+++ easy_sync/lib/pages/inputPage.dart
@@ -3,18 +3,22 @@ import 'dart:typed_data';
 
 import 'package:csv/csv.dart';
 import 'package:easy_sync/pages/editPage.dart';
+import 'package:easy_sync/pages/loginPage.dart';
 import 'package:easy_sync/tools/frame.dart';
+import 'package:easy_sync/tools/dataset.dart';
 import 'package:excel/excel.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 
 class InputPage extends StatefulWidget {
+  late Dataset table;
   _InputPageState createState() => _InputPageState();
 }
 
 class _InputPageState extends State<InputPage> {
   String? fileName;
   String? fileExtension;
+  bool useDefaults = true;
   List<List<String>> rows = [[]];
   List<String> columnNames = [];
 
@@ -95,10 +99,28 @@ class _InputPageState extends State<InputPage> {
 
   @override
   Widget build(BuildContext context) {
+    final Map<String, QuestionAndAnswers> questions = {
+      "summary": QuestionAndAnswers.withAnswers("How is each event title constructed", [0, 1, 7, 2]),
+      "location": QuestionAndAnswers.withAnswers("Where is the event Located", [19]),
+      "description": QuestionAndAnswers.withAnswers("While not needed. If you want to add a description, then you can build one here.", []),
+      "first day" : QuestionAndAnswers.withAnswers("Which column contains the first day", [14]),
+      "last day" : QuestionAndAnswers.withAnswers("Which column contains the last day", [14]),
+      "startTime" : QuestionAndAnswers.withAnswers("Which column contains the start time", [16]), 
+      "endTime" : QuestionAndAnswers.withAnswers("Which column contains the end time", [16]),
+      "recurrenceRules" : QuestionAndAnswers.withAnswers("Which column contains which days of the week are repeated", [16]),
+    };
     return Frame(
       title: 'Input',
       onNextPressed: () {
-        Navigator.push(context, MaterialPageRoute(builder: (context) => EditPage(rows)));
+        if (useDefaults) {
+          List<List<String>> newRows = rows.map((inner) => List<String>.from(inner)).toList();
+          newRows.removeAt(0);
+          widget.table = Dataset(rows);
+          
+          Navigator.push(context, MaterialPageRoute(builder: (context) => LoginPage(widget.table.getEvents(questions))));
+        } else {
+          Navigator.push(context, MaterialPageRoute(builder: (context) => EditPage(rows)));
+        }
       },
       prevColor: Colors.grey,
       child: Center(
@@ -106,6 +128,7 @@ class _InputPageState extends State<InputPage> {
           mainAxisAlignment: MainAxisAlignment.center,
           crossAxisAlignment: CrossAxisAlignment.center,
           children: [
+            if (fileName != null) const SizedBox(height: 50,),
             IconButton(
               onPressed: getFile, 
               icon: const Icon(Icons.file_open, color: Colors.blue, size: 100.0),
@@ -120,12 +143,22 @@ class _InputPageState extends State<InputPage> {
                     'File uploaded: $fileName',
                     textAlign: TextAlign.center,
                   ),
-                  // ElevatedButton(
-                  //   onPressed: () {
-                  //     Navigator.push(context, MaterialPageRoute(builder: (context) => Frame(title: 'edit', child: EditPage(rows))));
-                  //   }, 
-                  //   child: const Text("Edit data")
-                  // )
+                  const SizedBox(height: 35,),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      Checkbox(
+                        value: useDefaults, 
+                        onChanged: (bool? value) {
+                          setState(() {
+                            useDefaults = value!;
+                          });
+                          print("Use defaults: $useDefaults");
+                        }
+                      ),
+                      const Text("Use defaults")
+                    ],
+                  )
                 ],
               )
             ),
