+++ easy_sync/lib/tools/question_widget.dart
@@ -1,4 +1,5 @@
 import 'package:easy_sync/pages/editPage.dart';
+import 'package:easy_sync/tools/dataset.dart';
 import 'package:flutter/material.dart';
 
 class QuestionWidget extends StatelessWidget {
